Betas as hyperparams
